Create README.md
clickandeat
===========

ClickAndEat is an integrated application to support on-line takeaway food ordering. 

It includes an online administration tool where you can add restaurants along with their contact details, menus, opening times and any discounts they offer. The admin tool also includes an order summary view to enable billing and payments. 

The system supports geospatial querying to provide end users with a list of restaurants near to them. The system uses Twilio for automated phone calling and text messaging.

When a user places an order with ClickAndEat, the application places an automated call to the restaurant giving a summary of the order with the option to reject the order or hear the full details by pressing 1 or 2 on their phone keypad. If the restaurant accepts the order then the application reads out the full order details and lets them accept or reject the order.

The system also notifies the restaurant via email which includes html links to accept or reject the order.

The system notifies the end user whether their order has been accepted or rejected and if accepted, indicates how long the order will take to arrive.